requirements.txt
Flask==2.0.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Werkzeug==2.0.1
SQLAlchemy
web3==6.0.0
python-dotenv==0.19.0
email-validator==2.1.0.post1
